Build Nomulus with Java 17

+++ core/src/main/java/google/registry/model/tld/label/ReservedList.java
@@ -264,7 +264,7 @@ public final class ReservedList
     DomainLabelMetrics.recordReservedListCheckOutcome(
         tldStr,
         metricMatchesBuilder.build(),
-        DateTime.now(UTC).getMillis() - startTime.getMillis());
+        (double) (DateTime.now(UTC).getMillis() - startTime.getMillis()));
     return entries;
   }
